extract mock data into own file
