+++ packages/ui-components/src/FormBuilder/index.ts
@@ -1,2 +1,3 @@
 export { default as FormBuilder } from './FormBuilder';
+export * from './mockData';
 export * from './types';
